@@ -19,7 +19,7 @@ publish_to: "none"
 version: 1.0.0+1
 
 environment:
-    sdk: ">=2.17.5 <3.0.0"
+    sdk: ^3.8.1
 
 dependencies:
     flutter:
@@ -27,15 +27,22 @@ dependencies:
 
     # The following adds the Cupertino Icons font to your application.
     # Use with the CupertinoIcons class for iOS style icons.
-    cupertino_icons: ^1.0.2
-    font_awesome_flutter: ^10.1.0
+    cupertino_icons: ^1.0.8
+    font_awesome_flutter: ^10.8.0
     flutter_rating_bar: ^4.0.1
-    intl: ^0.17.0
-    animations: ^2.0.3
+    intl: ^0.20.2
+    animations: ^2.0.11
 
 dev_dependencies:
     flutter_test:
         sdk: flutter
+
+    # The "flutter_lints" package below contains a set of recommended lints to
+    # encourage good coding practices. The lint set provided by the package is
+    # activated in the `analysis_options.yaml` file located at the root of your
+    # package. See that file for information about deactivating specific lint
+    # rules and activating additional ones.
+    flutter_lints: ^6.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
