@@ -17,7 +17,9 @@ class _PopularCourseListViewState extends State<PopularCourseListView>
   @override
   void initState() {
     animationController = AnimationController(
-        duration: const Duration(milliseconds: 2000), vsync: this);
+      duration: const Duration(milliseconds: 2000),
+      vsync: this,
+    );
     super.initState();
   }
 
@@ -46,12 +48,15 @@ class _PopularCourseListViewState extends State<PopularCourseListView>
                   final int count = Category.popularCourseList.length;
                   final Animation<double> animation =
                       Tween<double>(begin: 0.0, end: 1.0).animate(
-                    CurvedAnimation(
-                      parent: animationController!,
-                      curve: Interval((1 / count) * index, 1.0,
-                          curve: Curves.fastOutSlowIn),
-                    ),
-                  );
+                        CurvedAnimation(
+                          parent: animationController!,
+                          curve: Interval(
+                            (1 / count) * index,
+                            1.0,
+                            curve: Curves.fastOutSlowIn,
+                          ),
+                        ),
+                      );
                   animationController?.forward();
                   return CategoryView(
                     callback: widget.callBack,
@@ -76,13 +81,13 @@ class _PopularCourseListViewState extends State<PopularCourseListView>
 }
 
 class CategoryView extends StatelessWidget {
-  const CategoryView(
-      {Key? key,
-      this.category,
-      this.animationController,
-      this.animation,
-      this.callback})
-      : super(key: key);
+  const CategoryView({
+    Key? key,
+    this.category,
+    this.animationController,
+    this.animation,
+    this.callback,
+  }) : super(key: key);
 
   final VoidCallback? callback;
   final Category? category;
@@ -98,7 +103,10 @@ class CategoryView extends StatelessWidget {
           opacity: animation!,
           child: Transform(
             transform: Matrix4.translationValues(
-                0.0, 50 * (1.0 - animation!.value), 0.0),
+              0.0,
+              50 * (1.0 - animation!.value),
+              0.0,
+            ),
             child: InkWell(
               splashColor: Colors.transparent,
               onTap: callback,
@@ -115,7 +123,8 @@ class CategoryView extends StatelessWidget {
                               decoration: BoxDecoration(
                                 color: HexColor('#F8FAFB'),
                                 borderRadius: const BorderRadius.all(
-                                    Radius.circular(16.0)),
+                                  Radius.circular(16.0),
+                                ),
                                 // border: new Border.all(
                                 //     color: DesignCourseAppTheme.notWhite),
                               ),
@@ -127,7 +136,10 @@ class CategoryView extends StatelessWidget {
                                         children: <Widget>[
                                           Padding(
                                             padding: const EdgeInsets.only(
-                                                top: 16, left: 16, right: 16),
+                                              top: 16,
+                                              left: 16,
+                                              right: 16,
+                                            ),
                                             child: Text(
                                               category!.title,
                                               textAlign: TextAlign.left,
@@ -142,10 +154,11 @@ class CategoryView extends StatelessWidget {
                                           ),
                                           Padding(
                                             padding: const EdgeInsets.only(
-                                                top: 8,
-                                                left: 16,
-                                                right: 16,
-                                                bottom: 8),
+                                              top: 0,
+                                              left: 16,
+                                              right: 16,
+                                              bottom: 8,
+                                            ),
                                             child: Row(
                                               mainAxisAlignment:
                                                   MainAxisAlignment
@@ -164,33 +177,30 @@ class CategoryView extends StatelessWidget {
                                                         .grey,
                                                   ),
                                                 ),
-                                                Container(
-                                                  child: Row(
-                                                    children: <Widget>[
-                                                      Text(
-                                                        '${category!.rating}',
-                                                        textAlign:
-                                                            TextAlign.left,
-                                                        style: TextStyle(
-                                                          fontWeight:
-                                                              FontWeight.w200,
-                                                          fontSize: 18,
-                                                          letterSpacing: 0.27,
-                                                          color:
-                                                              DesignCourseAppTheme
-                                                                  .grey,
-                                                        ),
-                                                      ),
-                                                      Icon(
-                                                        Icons.star,
+                                                Row(
+                                                  children: <Widget>[
+                                                    Text(
+                                                      '${category!.rating}',
+                                                      textAlign: TextAlign.left,
+                                                      style: TextStyle(
+                                                        fontWeight:
+                                                            FontWeight.w200,
+                                                        fontSize: 18,
+                                                        letterSpacing: 0.27,
                                                         color:
                                                             DesignCourseAppTheme
-                                                                .nearlyBlue,
-                                                        size: 20,
+                                                                .grey,
                                                       ),
-                                                    ],
-                                                  ),
-                                                )
+                                                    ),
+                                                    Icon(
+                                                      Icons.star,
+                                                      color:
+                                                          DesignCourseAppTheme
+                                                              .nearlyBlue,
+                                                      size: 20,
+                                                    ),
+                                                  ],
+                                                ),
                                               ],
                                             ),
                                           ),
@@ -198,41 +208,45 @@ class CategoryView extends StatelessWidget {
                                       ),
                                     ),
                                   ),
-                                  const SizedBox(
-                                    width: 48,
-                                  ),
+                                  const SizedBox(width: 48),
                                 ],
                               ),
                             ),
                           ),
-                          const SizedBox(
-                            height: 48,
-                          ),
+                          const SizedBox(height: 48),
                         ],
                       ),
                     ),
                     Container(
                       child: Padding(
-                        padding:
-                            const EdgeInsets.only(top: 24, right: 16, left: 16),
+                        padding: const EdgeInsets.only(
+                          top: 24,
+                          right: 16,
+                          left: 16,
+                        ),
                         child: Container(
                           decoration: BoxDecoration(
-                            borderRadius:
-                                const BorderRadius.all(Radius.circular(16.0)),
+                            borderRadius: const BorderRadius.all(
+                              Radius.circular(16.0),
+                            ),
                             boxShadow: <BoxShadow>[
                               BoxShadow(
-                                  color: DesignCourseAppTheme.grey
-                                      .withOpacity(0.2),
-                                  offset: const Offset(0.0, 0.0),
-                                  blurRadius: 6.0),
+                                color: DesignCourseAppTheme.grey.withOpacity(
+                                  0.2,
+                                ),
+                                offset: const Offset(0.0, 0.0),
+                                blurRadius: 6.0,
+                              ),
                             ],
                           ),
                           child: ClipRRect(
-                            borderRadius:
-                                const BorderRadius.all(Radius.circular(16.0)),
+                            borderRadius: const BorderRadius.all(
+                              Radius.circular(16.0),
+                            ),
                             child: AspectRatio(
-                                aspectRatio: 1.28,
-                                child: Image.asset(category!.imagePath)),
+                              aspectRatio: 1.28,
+                              child: Image.asset(category!.imagePath),
+                            ),
                           ),
                         ),
                       ),
