@@ -24,7 +24,9 @@ class _HotelHomeScreenState extends State<HotelHomeScreen>
   @override
   void initState() {
     animationController = AnimationController(
-        duration: const Duration(milliseconds: 1000), vsync: this);
+      duration: const Duration(milliseconds: 1000),
+      vsync: this,
+    );
     super.initState();
   }
 
@@ -63,44 +65,50 @@ class _HotelHomeScreenState extends State<HotelHomeScreen>
                         controller: _scrollController,
                         headerSliverBuilder:
                             (BuildContext context, bool innerBoxIsScrolled) {
-                          return <Widget>[
-                            SliverList(
-                              delegate: SliverChildBuilderDelegate(
-                                  (BuildContext context, int index) {
-                                return Column(
-                                  children: <Widget>[
-                                    getSearchBarUI(),
-                                    getTimeDateUI(),
-                                  ],
-                                );
-                              }, childCount: 1),
-                            ),
-                            SliverPersistentHeader(
-                              pinned: true,
-                              floating: true,
-                              delegate: ContestTabHeader(
-                                getFilterBarUI(),
-                              ),
-                            ),
-                          ];
-                        },
+                              return <Widget>[
+                                SliverList(
+                                  delegate: SliverChildBuilderDelegate((
+                                    BuildContext context,
+                                    int index,
+                                  ) {
+                                    return Column(
+                                      children: <Widget>[
+                                        getSearchBarUI(),
+                                        getTimeDateUI(),
+                                      ],
+                                    );
+                                  }, childCount: 1),
+                                ),
+                                SliverPersistentHeader(
+                                  pinned: true,
+                                  floating: true,
+                                  delegate: ContestTabHeader(getFilterBarUI()),
+                                ),
+                              ];
+                            },
                         body: Container(
-                          color:
-                              HotelAppTheme.buildLightTheme().backgroundColor,
+                          color: HotelAppTheme.buildLightTheme()
+                              .colorScheme
+                              .background,
                           child: ListView.builder(
                             itemCount: hotelList.length,
                             padding: const EdgeInsets.only(top: 8),
                             scrollDirection: Axis.vertical,
                             itemBuilder: (BuildContext context, int index) {
-                              final int count =
-                                  hotelList.length > 10 ? 10 : hotelList.length;
+                              final int count = hotelList.length > 10
+                                  ? 10
+                                  : hotelList.length;
                               final Animation<double> animation =
                                   Tween<double>(begin: 0.0, end: 1.0).animate(
-                                      CurvedAnimation(
-                                          parent: animationController!,
-                                          curve: Interval(
-                                              (1 / count) * index, 1.0,
-                                              curve: Curves.fastOutSlowIn)));
+                                    CurvedAnimation(
+                                      parent: animationController!,
+                                      curve: Interval(
+                                        (1 / count) * index,
+                                        1.0,
+                                        curve: Curves.fastOutSlowIn,
+                                      ),
+                                    ),
+                                  );
                               animationController?.forward();
                               return HotelListView(
                                 callback: () {},
@@ -112,7 +120,7 @@ class _HotelHomeScreenState extends State<HotelHomeScreen>
                           ),
                         ),
                       ),
-                    )
+                    ),
                   ],
                 ),
               ),
@@ -126,12 +134,13 @@ class _HotelHomeScreenState extends State<HotelHomeScreen>
   Widget getListUI() {
     return Container(
       decoration: BoxDecoration(
-        color: HotelAppTheme.buildLightTheme().backgroundColor,
+        color: HotelAppTheme.buildLightTheme().colorScheme.background,
         boxShadow: <BoxShadow>[
           BoxShadow(
-              color: Colors.grey.withOpacity(0.2),
-              offset: const Offset(0, -2),
-              blurRadius: 8.0),
+            color: Colors.grey.withOpacity(0.2),
+            offset: const Offset(0, -2),
+            blurRadius: 8.0,
+          ),
         ],
       ),
       child: Column(
@@ -148,14 +157,20 @@ class _HotelHomeScreenState extends State<HotelHomeScreen>
                     itemCount: hotelList.length,
                     scrollDirection: Axis.vertical,
                     itemBuilder: (BuildContext context, int index) {
-                      final int count =
-                          hotelList.length > 10 ? 10 : hotelList.length;
+                      final int count = hotelList.length > 10
+                          ? 10
+                          : hotelList.length;
                       final Animation<double> animation =
                           Tween<double>(begin: 0.0, end: 1.0).animate(
-                              CurvedAnimation(
-                                  parent: animationController!,
-                                  curve: Interval((1 / count) * index, 1.0,
-                                      curve: Curves.fastOutSlowIn)));
+                            CurvedAnimation(
+                              parent: animationController!,
+                              curve: Interval(
+                                (1 / count) * index,
+                                1.0,
+                                curve: Curves.fastOutSlowIn,
+                              ),
+                            ),
+                          );
                       animationController?.forward();
 
                       return HotelListView(
@@ -169,7 +184,7 @@ class _HotelHomeScreenState extends State<HotelHomeScreen>
                 }
               },
             ),
-          )
+          ),
         ],
       ),
     );
@@ -179,13 +194,17 @@ class _HotelHomeScreenState extends State<HotelHomeScreen>
     final List<Widget> hotelListViews = <Widget>[];
     for (int i = 0; i < hotelList.length; i++) {
       final int count = hotelList.length;
-      final Animation<double> animation =
-          Tween<double>(begin: 0.0, end: 1.0).animate(
-        CurvedAnimation(
-          parent: animationController!,
-          curve: Interval((1 / count) * i, 1.0, curve: Curves.fastOutSlowIn),
-        ),
-      );
+      final Animation<double> animation = Tween<double>(begin: 0.0, end: 1.0)
+          .animate(
+            CurvedAnimation(
+              parent: animationController!,
+              curve: Interval(
+                (1 / count) * i,
+                1.0,
+                curve: Curves.fastOutSlowIn,
+              ),
+            ),
+          );
       hotelListViews.add(
         HotelListView(
           callback: () {},
@@ -196,9 +215,7 @@ class _HotelHomeScreenState extends State<HotelHomeScreen>
       );
     }
     animationController?.forward();
-    return Column(
-      children: hotelListViews,
-    );
+    return Column(children: hotelListViews);
   }
 
   Widget getTimeDateUI() {
@@ -216,9 +233,7 @@ class _HotelHomeScreenState extends State<HotelHomeScreen>
                     highlightColor: Colors.transparent,
                     hoverColor: Colors.transparent,
                     splashColor: Colors.grey.withOpacity(0.2),
-                    borderRadius: const BorderRadius.all(
-                      Radius.circular(4.0),
-                    ),
+                    borderRadius: const BorderRadius.all(Radius.circular(4.0)),
                     onTap: () {
                       FocusScope.of(context).requestFocus(FocusNode());
                       // setState(() {
@@ -228,7 +243,11 @@ class _HotelHomeScreenState extends State<HotelHomeScreen>
                     },
                     child: Padding(
                       padding: const EdgeInsets.only(
-                          left: 8, right: 8, top: 4, bottom: 4),
+                        left: 8,
+                        right: 8,
+                        top: 4,
+                        bottom: 4,
+                      ),
                       child: Column(
                         mainAxisAlignment: MainAxisAlignment.center,
                         crossAxisAlignment: CrossAxisAlignment.start,
@@ -236,13 +255,12 @@ class _HotelHomeScreenState extends State<HotelHomeScreen>
                           Text(
                             'Choose date',
                             style: TextStyle(
-                                fontWeight: FontWeight.w100,
-                                fontSize: 16,
-                                color: Colors.grey.withOpacity(0.8)),
+                              fontWeight: FontWeight.w100,
+                              fontSize: 16,
+                              color: Colors.grey.withOpacity(0.8),
+                            ),
                           ),
-                          const SizedBox(
-                            height: 8,
-                          ),
+                          const SizedBox(height: 8),
                           Text(
                             '${DateFormat("dd, MMM").format(startDate)} - ${DateFormat("dd, MMM").format(endDate)}',
                             style: TextStyle(
@@ -276,15 +294,17 @@ class _HotelHomeScreenState extends State<HotelHomeScreen>
                     highlightColor: Colors.transparent,
                     hoverColor: Colors.transparent,
                     splashColor: Colors.grey.withOpacity(0.2),
-                    borderRadius: const BorderRadius.all(
-                      Radius.circular(4.0),
-                    ),
+                    borderRadius: const BorderRadius.all(Radius.circular(4.0)),
                     onTap: () {
                       FocusScope.of(context).requestFocus(FocusNode());
                     },
                     child: Padding(
                       padding: const EdgeInsets.only(
-                          left: 8, right: 8, top: 4, bottom: 4),
+                        left: 8,
+                        right: 8,
+                        top: 4,
+                        bottom: 4,
+                      ),
                       child: Column(
                         mainAxisAlignment: MainAxisAlignment.center,
                         crossAxisAlignment: CrossAxisAlignment.start,
@@ -292,13 +312,12 @@ class _HotelHomeScreenState extends State<HotelHomeScreen>
                           Text(
                             'Number of Rooms',
                             style: TextStyle(
-                                fontWeight: FontWeight.w100,
-                                fontSize: 16,
-                                color: Colors.grey.withOpacity(0.8)),
+                              fontWeight: FontWeight.w100,
+                              fontSize: 16,
+                              color: Colors.grey.withOpacity(0.8),
+                            ),
                           ),
-                          const SizedBox(
-                            height: 8,
-                          ),
+                          const SizedBox(height: 8),
                           Text(
                             '1 Room - 2 Adults',
                             style: TextStyle(
@@ -329,25 +348,26 @@ class _HotelHomeScreenState extends State<HotelHomeScreen>
               padding: const EdgeInsets.only(right: 16, top: 8, bottom: 8),
               child: Container(
                 decoration: BoxDecoration(
-                  color: HotelAppTheme.buildLightTheme().backgroundColor,
-                  borderRadius: const BorderRadius.all(
-                    Radius.circular(38.0),
-                  ),
+                  color: HotelAppTheme.buildLightTheme().colorScheme.background,
+                  borderRadius: const BorderRadius.all(Radius.circular(38.0)),
                   boxShadow: <BoxShadow>[
                     BoxShadow(
-                        color: Colors.grey.withOpacity(0.2),
-                        offset: const Offset(0, 2),
-                        blurRadius: 8.0),
+                      color: Colors.grey.withOpacity(0.2),
+                      offset: const Offset(0, 2),
+                      blurRadius: 8.0,
+                    ),
                   ],
                 ),
                 child: Padding(
                   padding: const EdgeInsets.only(
-                      left: 16, right: 16, top: 4, bottom: 4),
+                    left: 16,
+                    right: 16,
+                    top: 4,
+                    bottom: 4,
+                  ),
                   child: TextField(
                     onChanged: (String txt) {},
-                    style: const TextStyle(
-                      fontSize: 18,
-                    ),
+                    style: const TextStyle(fontSize: 18),
                     cursorColor: HotelAppTheme.buildLightTheme().primaryColor,
                     decoration: InputDecoration(
                       border: InputBorder.none,
@@ -361,30 +381,30 @@ class _HotelHomeScreenState extends State<HotelHomeScreen>
           Container(
             decoration: BoxDecoration(
               color: HotelAppTheme.buildLightTheme().primaryColor,
-              borderRadius: const BorderRadius.all(
-                Radius.circular(38.0),
-              ),
+              borderRadius: const BorderRadius.all(Radius.circular(38.0)),
               boxShadow: <BoxShadow>[
                 BoxShadow(
-                    color: Colors.grey.withOpacity(0.4),
-                    offset: const Offset(0, 2),
-                    blurRadius: 8.0),
+                  color: Colors.grey.withOpacity(0.4),
+                  offset: const Offset(0, 2),
+                  blurRadius: 8.0,
+                ),
               ],
             ),
             child: Material(
               color: Colors.transparent,
               child: InkWell(
-                borderRadius: const BorderRadius.all(
-                  Radius.circular(32.0),
-                ),
+                borderRadius: const BorderRadius.all(Radius.circular(32.0)),
                 onTap: () {
                   FocusScope.of(context).requestFocus(FocusNode());
                 },
                 child: Padding(
                   padding: const EdgeInsets.all(16.0),
-                  child: Icon(FontAwesomeIcons.magnifyingGlass,
-                      size: 20,
-                      color: HotelAppTheme.buildLightTheme().backgroundColor),
+                  child: Icon(
+                    FontAwesomeIcons.magnifyingGlass,
+                    size: 20,
+                    color:
+                        HotelAppTheme.buildLightTheme().colorScheme.background,
+                  ),
                 ),
               ),
             ),
@@ -404,21 +424,26 @@ class _HotelHomeScreenState extends State<HotelHomeScreen>
           child: Container(
             height: 24,
             decoration: BoxDecoration(
-              color: HotelAppTheme.buildLightTheme().backgroundColor,
+              color: HotelAppTheme.buildLightTheme().colorScheme.background,
               boxShadow: <BoxShadow>[
                 BoxShadow(
-                    color: Colors.grey.withOpacity(0.2),
-                    offset: const Offset(0, -2),
-                    blurRadius: 8.0),
+                  color: Colors.grey.withOpacity(0.2),
+                  offset: const Offset(0, -2),
+                  blurRadius: 8.0,
+                ),
               ],
             ),
           ),
         ),
         Container(
-          color: HotelAppTheme.buildLightTheme().backgroundColor,
+          color: HotelAppTheme.buildLightTheme().colorScheme.background,
           child: Padding(
-            padding:
-                const EdgeInsets.only(left: 16, right: 16, top: 8, bottom: 4),
+            padding: const EdgeInsets.only(
+              left: 16,
+              right: 16,
+              top: 8,
+              bottom: 4,
+            ),
             child: Row(
               children: <Widget>[
                 Expanded(
@@ -440,16 +465,15 @@ class _HotelHomeScreenState extends State<HotelHomeScreen>
                     highlightColor: Colors.transparent,
                     hoverColor: Colors.transparent,
                     splashColor: Colors.grey.withOpacity(0.2),
-                    borderRadius: const BorderRadius.all(
-                      Radius.circular(4.0),
-                    ),
+                    borderRadius: const BorderRadius.all(Radius.circular(4.0)),
                     onTap: () {
                       FocusScope.of(context).requestFocus(FocusNode());
                       Navigator.push<dynamic>(
                         context,
                         MaterialPageRoute<dynamic>(
-                            builder: (BuildContext context) => FiltersScreen(),
-                            fullscreenDialog: true),
+                          builder: (BuildContext context) => FiltersScreen(),
+                          fullscreenDialog: true,
+                        ),
                       );
                     },
                     child: Padding(
@@ -465,9 +489,11 @@ class _HotelHomeScreenState extends State<HotelHomeScreen>
                           ),
                           Padding(
                             padding: const EdgeInsets.all(8.0),
-                            child: Icon(Icons.sort,
-                                color: HotelAppTheme.buildLightTheme()
-                                    .primaryColor),
+                            child: Icon(
+                              Icons.sort,
+                              color:
+                                  HotelAppTheme.buildLightTheme().primaryColor,
+                            ),
                           ),
                         ],
                       ),
@@ -478,14 +504,7 @@ class _HotelHomeScreenState extends State<HotelHomeScreen>
             ),
           ),
         ),
-        const Positioned(
-          top: 0,
-          left: 0,
-          right: 0,
-          child: Divider(
-            height: 1,
-          ),
-        )
+        const Positioned(top: 0, left: 0, right: 0, child: Divider(height: 1)),
       ],
     );
   }
@@ -513,17 +532,21 @@ class _HotelHomeScreenState extends State<HotelHomeScreen>
   Widget getAppBarUI() {
     return Container(
       decoration: BoxDecoration(
-        color: HotelAppTheme.buildLightTheme().backgroundColor,
+        color: HotelAppTheme.buildLightTheme().colorScheme.background,
         boxShadow: <BoxShadow>[
           BoxShadow(
-              color: Colors.grey.withOpacity(0.2),
-              offset: const Offset(0, 2),
-              blurRadius: 8.0),
+            color: Colors.grey.withOpacity(0.2),
+            offset: const Offset(0, 2),
+            blurRadius: 8.0,
+          ),
         ],
       ),
       child: Padding(
         padding: EdgeInsets.only(
-            top: MediaQuery.of(context).padding.top, left: 8, right: 8),
+          top: MediaQuery.of(context).padding.top,
+          left: 8,
+          right: 8,
+        ),
         child: Row(
           children: <Widget>[
             Container(
@@ -533,9 +556,7 @@ class _HotelHomeScreenState extends State<HotelHomeScreen>
               child: Material(
                 color: Colors.transparent,
                 child: InkWell(
-                  borderRadius: const BorderRadius.all(
-                    Radius.circular(32.0),
-                  ),
+                  borderRadius: const BorderRadius.all(Radius.circular(32.0)),
                   onTap: () {
                     Navigator.pop(context);
                   },
@@ -550,10 +571,7 @@ class _HotelHomeScreenState extends State<HotelHomeScreen>
               child: Center(
                 child: Text(
                   'Explore',
-                  style: TextStyle(
-                    fontWeight: FontWeight.w600,
-                    fontSize: 22,
-                  ),
+                  style: TextStyle(fontWeight: FontWeight.w600, fontSize: 22),
                 ),
               ),
             ),
@@ -592,7 +610,7 @@ class _HotelHomeScreenState extends State<HotelHomeScreen>
                   ),
                 ],
               ),
-            )
+            ),
           ],
         ),
       ),
@@ -601,14 +619,15 @@ class _HotelHomeScreenState extends State<HotelHomeScreen>
 }
 
 class ContestTabHeader extends SliverPersistentHeaderDelegate {
-  ContestTabHeader(
-    this.searchUI,
-  );
+  ContestTabHeader(this.searchUI);
   final Widget searchUI;
 
   @override
   Widget build(
-      BuildContext context, double shrinkOffset, bool overlapsContent) {
+    BuildContext context,
+    double shrinkOffset,
+    bool overlapsContent,
+  ) {
     return searchUI;
   }
 
