@@ -4,71 +4,53 @@ class CareView extends StatelessWidget {
   final AnimationController animationController;
 
   const CareView({Key? key, required this.animationController})
-      : super(key: key);
+    : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     final _firstHalfAnimation =
-        Tween<Offset>(begin: Offset(1, 0), end: Offset(0, 0))
-            .animate(CurvedAnimation(
-      parent: animationController,
-      curve: Interval(
-        0.2,
-        0.4,
-        curve: Curves.fastOutSlowIn,
-      ),
-    ));
+        Tween<Offset>(begin: Offset(1, 0), end: Offset(0, 0)).animate(
+          CurvedAnimation(
+            parent: animationController,
+            curve: Interval(0.2, 0.4, curve: Curves.fastOutSlowIn),
+          ),
+        );
     final _secondHalfAnimation =
-        Tween<Offset>(begin: Offset(0, 0), end: Offset(-1, 0))
-            .animate(CurvedAnimation(
-      parent: animationController,
-      curve: Interval(
-        0.4,
-        0.6,
-        curve: Curves.fastOutSlowIn,
-      ),
-    ));
+        Tween<Offset>(begin: Offset(0, 0), end: Offset(-1, 0)).animate(
+          CurvedAnimation(
+            parent: animationController,
+            curve: Interval(0.4, 0.6, curve: Curves.fastOutSlowIn),
+          ),
+        );
     final _relaxFirstHalfAnimation =
-        Tween<Offset>(begin: Offset(2, 0), end: Offset(0, 0))
-            .animate(CurvedAnimation(
-      parent: animationController,
-      curve: Interval(
-        0.2,
-        0.4,
-        curve: Curves.fastOutSlowIn,
-      ),
-    ));
+        Tween<Offset>(begin: Offset(2, 0), end: Offset(0, 0)).animate(
+          CurvedAnimation(
+            parent: animationController,
+            curve: Interval(0.2, 0.4, curve: Curves.fastOutSlowIn),
+          ),
+        );
     final _relaxSecondHalfAnimation =
-        Tween<Offset>(begin: Offset(0, 0), end: Offset(-2, 0))
-            .animate(CurvedAnimation(
-      parent: animationController,
-      curve: Interval(
-        0.4,
-        0.6,
-        curve: Curves.fastOutSlowIn,
-      ),
-    ));
+        Tween<Offset>(begin: Offset(0, 0), end: Offset(-2, 0)).animate(
+          CurvedAnimation(
+            parent: animationController,
+            curve: Interval(0.4, 0.6, curve: Curves.fastOutSlowIn),
+          ),
+        );
 
     final _imageFirstHalfAnimation =
-        Tween<Offset>(begin: Offset(4, 0), end: Offset(0, 0))
-            .animate(CurvedAnimation(
-      parent: animationController,
-      curve: Interval(
-        0.2,
-        0.4,
-        curve: Curves.fastOutSlowIn,
-      ),
-    ));
+        Tween<Offset>(begin: Offset(4, 0), end: Offset(0, 0)).animate(
+          CurvedAnimation(
+            parent: animationController,
+            curve: Interval(0.2, 0.4, curve: Curves.fastOutSlowIn),
+          ),
+        );
     final _imageSecondHalfAnimation =
-        Tween<Offset>(begin: Offset(0, 0), end: Offset(-4, 0))
-            .animate(CurvedAnimation(
-      parent: animationController,
-      curve: Interval(
-        0.4,
-        0.6,
-        curve: Curves.fastOutSlowIn,
-      ),
-    ));
+        Tween<Offset>(begin: Offset(0, 0), end: Offset(-4, 0)).animate(
+          CurvedAnimation(
+            parent: animationController,
+            curve: Interval(0.4, 0.6, curve: Curves.fastOutSlowIn),
+          ),
+        );
 
     return SlideTransition(
       position: _firstHalfAnimation,
@@ -98,14 +80,20 @@ class CareView extends StatelessWidget {
                   position: _relaxSecondHalfAnimation,
                   child: Text(
                     "Care",
-                    style:
-                        TextStyle(fontSize: 26.0, fontWeight: FontWeight.bold),
+                    style: TextStyle(
+                      fontSize: 26.0,
+                      fontWeight: FontWeight.bold,
+                    ),
                   ),
                 ),
               ),
               Padding(
-                padding:
-                    EdgeInsets.only(left: 64, right: 64, bottom: 16, top: 16),
+                padding: EdgeInsets.only(
+                  left: 64,
+                  right: 64,
+                  bottom: 16,
+                  top: 0,
+                ),
                 child: Text(
                   "Lorem ipsum dolor sit amet,consectetur adipiscing elit,sed do eiusmod tempor incididunt ut labore",
                   textAlign: TextAlign.center,
