@@ -4,7 +4,7 @@ class SplashView extends StatefulWidget {
   final AnimationController animationController;
 
   const SplashView({Key? key, required this.animationController})
-      : super(key: key);
+    : super(key: key);
 
   @override
   _SplashViewState createState() => _SplashViewState();
@@ -14,15 +14,12 @@ class _SplashViewState extends State<SplashView> {
   @override
   Widget build(BuildContext context) {
     final _introductionanimation =
-        Tween<Offset>(begin: Offset(0, 0), end: Offset(0.0, -1.0))
-            .animate(CurvedAnimation(
-      parent: widget.animationController,
-      curve: Interval(
-        0.0,
-        0.2,
-        curve: Curves.fastOutSlowIn,
-      ),
-    ));
+        Tween<Offset>(begin: Offset(0, 0), end: Offset(0.0, -1.0)).animate(
+          CurvedAnimation(
+            parent: widget.animationController,
+            curve: Interval(0.0, 0.2, curve: Curves.fastOutSlowIn),
+          ),
+        );
     return SlideTransition(
       position: _introductionanimation,
       child: SingleChildScrollView(
@@ -38,7 +35,7 @@ class _SplashViewState extends State<SplashView> {
             Padding(
               padding: EdgeInsets.only(top: 8.0, bottom: 8.0),
               child: Text(
-                "Clearhead",
+                "ClearHead",
                 style: TextStyle(fontSize: 25.0, fontWeight: FontWeight.bold),
               ),
             ),
@@ -49,12 +46,11 @@ class _SplashViewState extends State<SplashView> {
                 textAlign: TextAlign.center,
               ),
             ),
-            SizedBox(
-              height: 48,
-            ),
+            SizedBox(height: 48),
             Padding(
               padding: EdgeInsets.only(
-                  bottom: MediaQuery.of(context).padding.bottom + 16),
+                bottom: MediaQuery.of(context).padding.bottom + 16,
+              ),
               child: InkWell(
                 onTap: () {
                   widget.animationController.animateTo(0.2);
@@ -73,10 +69,7 @@ class _SplashViewState extends State<SplashView> {
                   ),
                   child: Text(
                     "Let's begin",
-                    style: TextStyle(
-                      fontSize: 18,
-                      color: Colors.white,
-                    ),
+                    style: TextStyle(fontSize: 18, color: Colors.white),
                   ),
                 ),
               ),
